# lib/api-zod/src/generated/types/discordIdentity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiscordIdentity {
  id: string;
  username: string;
  /** @nullable */
  globalName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatar?: string | null;
}