# lib/api-zod/src/generated/types/guildChannel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GuildChannel {
  id: string;
  name: string;
  type: number;
  position: number;
}