# lib/api-zod/src/generated/types/guildSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GuildSummary {
  id: string;
  guildId: string;
  welcomeEnabled: boolean;
  partnerEnabled: boolean;
  hasTicketChannel: boolean;
  hasShop: boolean;
  isOwner: boolean;
  permissions: number;
  /** @nullable */
  memberCount?: number | null;
  /** @nullable */
  discordName?: string | null;
  /** @nullable */
  discordIcon?: string | null;
}